Update Crowd-Fund-Tutorial-with-Eth-Brownie-on-Celo.md
-	In the modified code, I have added the following error handling:

a)	In the safeMint function, I added a check to ensure that the connected account is eligible to mint a token. If the condition fails, it will throw an error.
b)	In the donate function, I added a check to ensure that donations cannot be made by the contract owner. If the condition fails, it will throw an error. Additionally, I added a check to ensure that the sent value is not below the minimum donation amount. If the condition fails, it will throw an error.
c)	In the transferDonations function, I added a check to ensure that the transfer is not made to the contract owner's account. If the condition fails, it will throw an error. I also added a check to ensure that the withdrawal limit is not exceeded. If the condition fails, it will throw an error.
d)	In the newPost function, I added a check to ensure that the post is not empty. If the condition fails, it will throw an error.
e)	In the likeAndUnlikePost function, I added a check to ensure that the post being liked or unliked exists. If the condition fails, it will throw an error.
f)	In the deletePost function, I added a check to ensure that the account deleting the post is authorized to do so. If the condition fails, it will throw an error.

-	Typo in the _donationtBalance function: The function _donationtBalance has a typo in its name. It should be _donationBalance instead.
-	Duplicate "using Counters for Counters.Counter" statements: There are four duplicate "using Counters for Counters.Counter" statements at the beginning of the contract. Only one is necessary, so you can remove the extra three.
-	Unused variable _ActPstCount: The variable _ActPstCount is declared but not used in the code. If it's not needed, you can remove it to avoid confusion.
-	Added a section called “Entire code”.
-	Added relevant comments to the code as they provide a brief overview of what the contract does and explain the purpose of each function and variable.
-	Applied code indentation so that it looks neat and aligned.